requirements.txt
psycopg2-binary
sqlalchemy
python-dotenv
selenium
webdriver-manager
python-dateutil
requests
pandas
time
datetime

